finish the slider Component
