+++ components/slider/slider.jsx
@@ -1,44 +1,77 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import RcSlider from 'rc-slider/lib/Slider';
+import RcRange from 'rc-slider/lib/Range';
+import RcHandle from 'rc-slider/lib/Handle';
+import Tooltip from '../tooltip';
 
-import React, {Component} from 'react';
-import RcSlider from 'rc-slider';
-// import Tooltip from '../tooltip';
-import './style/index';
+import './style'
 
-
-
-export default class Slider extends Component {
+export default class Slider extends React.Component {
+  static propTypes = {
+    min: PropTypes.number,
+    max: PropTypes.number,
+    marks: PropTypes.object
+  }
   static defaultProps = {
     prefixCls: 'idoll-slider',
-    tipTransitionName: 'zoom-down'
+    tooltipPrefixCls: 'idoll-tooltip',
+    tipFormatter(value) {
+      return value.toString();
+    },
+  }
+  constructor(props) {
+    super(props);
+    this.state = {
+      visibles: {},
+    }
+  }
+  toggleTooltipVisible = (index, visible) => {
+    this.setState(({visibles}) => ({
+      visibles: {
+        ...visibles,
+        [index]: visible
+      }
+    }));
+  }
+  handleWithTooltip = ({value, dragging, index, ...restProps}) => {
+    const { tooltipPrefixCls, tipFormatter } = this.props;
+    const { visibles } = this.state;
+    const visible = tipFormatter ? (visibles[index] || dragging) : false;
+    return (
+      <Tooltip
+        prefixCls={tooltipPrefixCls}
+        title={tipFormatter ? tipFormatter(value) : ''}
+        visible={visible}
+        placement='top'
+        key={index}
+      >
+        <RcHandle
+          {...restProps}
+          value={value}
+          onMouseEnter={() => this.toggleTooltipVisible(index, true)}
+          onMouseLeave={() => this.toggleTooltipVisible(index, false)}
+        />
+      </Tooltip>
+    );
+  }
+  focus() {
+    this.rcSlider.focus();
+  }
+
+  blur() {
+    this.rcSlider.focus();
+  }
+
+  saveSlider = (node) => {
+    this.rcSlider = node;
   }
 
   render() {
-    const { isIncluded, marks, index, defaultIndex, ...rest } = this.props;
-    if (isIncluded !== undefined) {
-      rest.included = isIncluded;
+    const { range, ...restProps } = this.props;
+    if (range) {
+      return <RcRange {...restProps} ref={this.saveSlider} handle={this.handleWithTooltip} />
     }
-    if (Array.isArray(marks)) {
-      rest.min = 0;
-      rest.max = marks.length - 1;
-      rest.step = 1;
-      if (index !== undefined) {
-        rest.value = index;
-      }
-      if (defaultIndex !== undefined) {
-        rest.defaultValue = defaultIndex;
-      }
-      rest.marks = {};
-      marks.forEach((val, idx) => {
-        rest.marks[idx] = val;
-      });
-    } else {
-      rest.marks = marks;
-    }
-
-    return (
-      <RcSlider {...rest} />
-    )
+    return <RcSlider {...restProps} ref={this.saveSlider} handle={this.handleWithTooltip} />
   }
 }
-
-
