--- components/slider/style/index.css
@@ -1,170 +1,0 @@
-
-.idoll-slider {
-  font-family: "Helvetica Neue For Number", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial, sans-serif;
-  font-size: 14px;
-  font-variant: tabular-nums;
-  line-height: 1.5;
-  color: rgba(0, 0, 0, 0.65);
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  position: relative;
-  margin: 14px 6px 10px;
-  padding: 4px 0;
-  height: 16px;
-  cursor: pointer;
-  touch-action: none;
-}
-.idoll-slider-vertical {
-  width: 12px;
-  height: 100%;
-  margin: 6px 10px;
-  padding: 0 4px;
-}
-.idoll-slider-vertical .idoll-slider-rail {
-  height: 100%;
-  width: 4px;
-}
-.idoll-slider-vertical .idoll-slider-track {
-  width: 4px;
-}
-.idoll-slider-vertical .idoll-slider-handle {
-  margin-left: -5px;
-  margin-bottom: -7px;
-}
-.idoll-slider-vertical .idoll-slider-mark {
-  top: 0;
-  left: 12px;
-  width: 18px;
-  height: 100%;
-}
-.idoll-slider-vertical .idoll-slider-mark-text {
-  left: 4px;
-  white-space: nowrap;
-}
-.idoll-slider-vertical .idoll-slider-step {
-  width: 4px;
-  height: 100%;
-}
-.idoll-slider-vertical .idoll-slider-dot {
-  top: auto;
-  left: 2px;
-  margin-bottom: -4px;
-}
-.idoll-slider-with-marks {
-  margin-bottom: 28px;
-}
-.idoll-slider-rail {
-  position: absolute;
-  width: 100%;
-  height: 4px;
-  border-radius: 2px;
-  background-color: #f5f5f5;
-  transition: background-color 0.3s;
-}
-.idoll-slider-track {
-  position: absolute;
-  height: 4px;
-  border-radius: 3px;
-  background-color: #81deb9;
-  transition: background-color 0.3s ease;
-}
-.idoll-slider-handle {
-  position: absolute;
-  margin-left: -8px;
-  margin-top: -6px;
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
-  border-radius: 50%;
-  border: solid 2px #81deb9;
-  background-color: #fff;
-  transition: border-color 0.3s, transform 0.3s cubic-bezier(0.18, 0.89, 0.32, 1.28);
-}
-.idoll-slider-handle:focus {
-  border-color: #42c69e;
-  height: 16px;
-  width: 16px;
-  margin-top: -6px;
-  box-shadow: 0 0 1px 6px #81deb9;
-  border: 2px solid #81deb9;
-  outline: none;
-}
-.idoll-slider-handle.idoll-tooltip-open {
-  border-color: #13B886;
-}
-.idoll-slider:hover .idoll-slider-rail {
-  background-color: #e1e1e1;
-}
-.idoll-slider:hover .idoll-slider-track {
-  background-color: #58d1a5;
-}
-.idoll-slider:hover .idoll-slider-handle:not(.idoll-tooltip-open) {
-  border-color: #58d1a5;
-  height: 16px;
-  width: 16px;
-  margin-top: -6px;
-}
-.idoll-slider-mark {
-  position: absolute;
-  top: 14px;
-  left: 0;
-  width: 100%;
-  font-size: 14px;
-}
-.idoll-slider-mark-text {
-  position: absolute;
-  display: inline-block;
-  vertical-align: middle;
-  text-align: center;
-  cursor: pointer;
-  color: rgba(0, 0, 0, 0.45);
-}
-.idoll-slider-mark-text-active {
-  color: rgba(0, 0, 0, 0.65);
-}
-.idoll-slider-step {
-  position: absolute;
-  width: 100%;
-  height: 4px;
-  background: transparent;
-}
-.idoll-slider-dot {
-  position: absolute;
-  top: -2px;
-  margin-left: -4px;
-  width: 8px;
-  height: 8px;
-  border: 2px solid #e8e8e8;
-  background-color: #fff;
-  cursor: pointer;
-  border-radius: 50%;
-  vertical-align: middle;
-}
-.idoll-slider-dot:first-child {
-  margin-left: -4px;
-}
-.idoll-slider-dot:last-child {
-  margin-left: -4px;
-}
-.idoll-slider-dot-active {
-  border-color: #89dcc3;
-}
-.idoll-slider-disabled {
-  cursor: not-allowed;
-}
-.idoll-slider-disabled .idoll-slider-track {
-  background-color: rgba(0, 0, 0, 0.25) !important;
-}
-.idoll-slider-disabled .idoll-slider-handle,
-.idoll-slider-disabled .idoll-slider-dot {
-  border-color: rgba(0, 0, 0, 0.25) !important;
-  background-color: #fff;
-  cursor: not-allowed;
-  box-shadow: none;
-}
-.idoll-slider-disabled .idoll-slider-mark-text,
-.idoll-slider-disabled .idoll-slider-dot {
-  cursor: not-allowed !important;
-}
